pyproject.toml
[tool.poetry]
name = "bukichi-bot"
version = "0.1.0"
description = "Discord bukichi bot"
authors = ["eddie-splatoonn <eddie.splatoon@gmail.com>"]
readme = "README.md"
packages = [{include = "bukichi_bot"}]

[tool.poetry.dependencies]
python = "^3.12"
discord-py = {extras = ["voice"], version = "^2.3.2"}
pillow = "^10.1.0"
regex = "^2023.12.25"
requests = "^2.31.0"
jupyter = "^1.0.0"
pydub = "^0.25.1"


[tool.poetry.group.dev.dependencies]
pytest = "^8.2.0"
pytest-cov = "^5.0.0"
pytest-asyncio = "^0.23.6"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
